backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.22
alembic==1.12.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
tensorflow==2.14.0
matplotlib==3.8.0
alpaca-trade-api==3.0.2
yfinance==0.2.28
pytest==7.4.2
python-dotenv==1.0.0
redis==5.0.1
influxdb-client==1.36.1
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4